move status from ingest common to rdw common (#205)
