+++ build.gradle
@@ -50,6 +50,7 @@ subprojects {
             dependency 'org.opentestsystem.rdw.common:rdw-common-archive:0.0.1-SNAPSHOT'
             dependency 'org.opentestsystem.rdw.common:rdw-common-messaging:0.0.1-SNAPSHOT'
             dependency 'org.opentestsystem.rdw.common:rdw-common-model:0.0.1-SNAPSHOT'
+            dependency 'org.opentestsystem.rdw.common:rdw-common-status:0.0.1-SNAPSHOT'
             dependency 'org.opentestsystem.rdw.common:rdw-common-utils:0.0.1-SNAPSHOT'
 
             dependency 'com.google.guava:guava:21.0'
