+++ common/build.gradle
@@ -16,11 +16,10 @@ dependencies {
     // handy for creating metadata with @ConfigurationProperties
     optional 'org.springframework.boot:spring-boot-configuration-processor'
 
-    // For StatusMvcEndpoint
-    compile 'org.springframework.boot:spring-boot-starter-actuator'
-    compile 'org.springframework.boot:spring-boot-starter-web'
+    // for @NotNull
+    compile 'javax.validation:validation-api'
 
-    compile "com.google.code.gson:gson:2.8.0"
+    compile 'com.google.code.gson:gson:2.8.0'
 
     testCompile 'org.springframework:spring-test'
     testCompile 'org.springframework.boot:spring-boot-starter-test'
