--- common/src/main/java/org/opentestsystem/rdw/ingest/common/status/ConfigurationStatusIndicator.java
@@ -1,35 +1,0 @@
-package org.opentestsystem.rdw.ingest.common.status;
-
-import org.springframework.boot.actuate.endpoint.ConfigurationPropertiesReportEndpoint;
-import org.springframework.boot.actuate.endpoint.EnvironmentEndpoint;
-
-/**
- * {@link StatusIndicator} for level 2 configuration diagnostics.
- */
-public class ConfigurationStatusIndicator extends AbstractStatusIndicator {
-
-    private final ConfigurationPropertiesReportEndpoint configurationEndpoint;
-    private final EnvironmentEndpoint environmentEndpoint;
-
-    public ConfigurationStatusIndicator(final ConfigurationPropertiesReportEndpoint configurationEndpoint,
-                                        final EnvironmentEndpoint environmentEndpoint) {
-        this.configurationEndpoint = configurationEndpoint;
-        this.environmentEndpoint = environmentEndpoint;
-    }
-
-    @Override
-    public String name() {
-        return "configuration";
-    }
-
-    @Override
-    protected boolean doLevelCheck(final int level) {
-        return level >= 2;
-    }
-
-    @Override
-    protected void doStatusCheck(final Status.Builder builder, final int level) {
-        builder.detail("environmentSettings", environmentEndpoint.invoke());
-        builder.detail("configClientProperties", configurationEndpoint.invoke().get("configClientProperties"));
-    }
-}
