--- common/src/main/java/org/opentestsystem/rdw/ingest/common/status/DiagnosticLevel.java
@@ -1,24 +1,0 @@
-package org.opentestsystem.rdw.ingest.common.status;
-
-public enum DiagnosticLevel {
-    Ping(0),
-    LocalSystem(1),
-    Configuration(2),
-    ReadDatabase(3),
-    WriteDatabase(4),
-    ExternalDependencies(5);
-
-    private final int value;
-
-    DiagnosticLevel(final int value) {
-        this.value = value;
-    }
-
-    /**
-     * @param level level to compare against
-     * @return true if this enum's value is less-than-or-equal-to the level
-     */
-    public boolean isLessThanOrEqualTo(final int level) {
-        return value <= level;
-    }
-}
