--- common/src/main/java/org/opentestsystem/rdw/ingest/common/status/Rating.java
@@ -1,22 +1,0 @@
-package org.opentestsystem.rdw.ingest.common.status;
-
-/**
- * Status rating defined by web diagnostic spec.
- */
-public enum Rating {
-    Failed(0),
-    Degraded(1),
-    Warning(2),
-    Recovering(3),
-    Ideal(4);
-
-    private final int value;
-
-    Rating(int value) {
-        this.value = value;
-    }
-
-    public int value() {
-        return value;
-    }
-}
