--- common/src/main/java/org/opentestsystem/rdw/ingest/common/status/StatusConfiguration.java
@@ -1,49 +1,0 @@
-package org.opentestsystem.rdw.ingest.common.status;
-
-import org.springframework.boot.actuate.endpoint.ConfigurationPropertiesReportEndpoint;
-import org.springframework.boot.actuate.endpoint.EnvironmentEndpoint;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import java.util.List;
-
-/**
- * {@link Configuration} for setting up default {@link StatusIndicator}s.
- * <p>
- * To add additional status information, simply create a @Bean/@Component that implements {@link StatusIndicator}
- * (extending {@link AbstractStatusIndicator} will take care of basic exception handling). StatusEndpoint will
- * integrate the information based on the request level. Please follow the web diagnostic specifications from
- * SmarterBalanced, http://www.smarterapp.org/documents/DiagnosticApi.html.
- * </p>
- * <p>
- * TODO - think about ProvidersIndicator; should we add additional framework so that we can provide a default
- * TODO   implementation that checks config-server, etc. while allowing extensibility?
- * </p>
- * <p>
- * NOTE: currently this just exposes the MVC bean. When it makes sense to directly expose StatusEndpoint
- * (e.g. if there is need for a JMX wrapper) do so.
- * Look at EndpointAutoConfiguration vs. EndpointWebMvcManagementContextConfiguration for a pattern to follow.
- * </p>
- *
- * @see StatusEndpoint
- */
-@Configuration
-public class StatusConfiguration {
-
-    @Bean
-    public StatusMvcEndpoint statusMvcEndpoint(final List<StatusIndicator> statusIndicators) {
-        return new StatusMvcEndpoint(new StatusEndpoint(statusIndicators));
-    }
-
-    @Bean
-    public LocalSystemStatusIndicator localSystemStatusIndicator() {
-        return new LocalSystemStatusIndicator();
-    }
-
-    @Bean
-    public ConfigurationStatusIndicator configurationStatusIndicator(
-            final ConfigurationPropertiesReportEndpoint configurationEndpoint,
-            final EnvironmentEndpoint environmentEndpoint) {
-        return new ConfigurationStatusIndicator(configurationEndpoint, environmentEndpoint);
-    }
-}
