--- common/src/main/java/org/opentestsystem/rdw/ingest/common/status/StatusEndpoint.java
@@ -1,44 +1,0 @@
-package org.opentestsystem.rdw.ingest.common.status;
-
-import org.springframework.boot.actuate.endpoint.AbstractEndpoint;
-import org.springframework.boot.actuate.endpoint.Endpoint;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-import java.util.List;
-
-/**
- * An {@link Endpoint} to expose web diagnostics.
- *
- * @see StatusConfiguration
- */
-@ConfigurationProperties(prefix = "endpoints.status")
-public class StatusEndpoint extends AbstractEndpoint<SystemStatus> {
-
-    private final List<StatusIndicator> statusIndicators;
-
-    public StatusEndpoint(final List<StatusIndicator> statusIndicators) {
-        super("status", true);
-        this.statusIndicators = statusIndicators;
-    }
-
-    @Override
-    public SystemStatus invoke() {
-        return invoke(0);
-    }
-
-    public SystemStatus invoke(final int level) {
-        return generateStatus(level);
-    }
-
-    private SystemStatus generateStatus(final int level) {
-        final SystemStatus.Builder builder = SystemStatus.builder()
-                .level(level);
-        for (final StatusIndicator statusIndicator : statusIndicators) {
-            final Status status = statusIndicator.status(level);
-            if (status != null) {
-                builder.sub(statusIndicator.name(), status);
-            }
-        }
-        return builder.build();
-    }
-}
