--- common/src/main/java/org/opentestsystem/rdw/ingest/common/status/StatusIndicator.java
@@ -1,21 +1,0 @@
-package org.opentestsystem.rdw.ingest.common.status;
-
-/**
- * A StatusIndicator produces a named status. The SystemStatus uses these indicators and the
- * requested diagnostic level to build an appropriate response.
- */
-public interface StatusIndicator {
-
-    /**
-     * @return name of status, e.g. "configuration"
-     */
-    String name();
-
-    /**
-     * Generate status for the given level
-     *
-     * @param level diagnostic level
-     * @return status for the given level, may be null
-     */
-    Status status(int level);
-}
