--- common/src/main/java/org/opentestsystem/rdw/ingest/common/status/UnitStatus.java
@@ -1,102 +1,0 @@
-package org.opentestsystem.rdw.ingest.common.status;
-
-import java.time.Instant;
-import java.util.Date;
-
-/**
- * The simplest status, it extends the base to include a unit name and status level.
- */
-public class UnitStatus extends BaseStatus {
-    private String unit;
-    private int level;
-    private String message;
-    private Date dateTime;
-
-    public String getUnit() {
-        return unit;
-    }
-
-    void setUnit(final String unit) {
-        this.unit = unit;
-    }
-
-    /**
-     * @return diagnostic status level, 0-5
-     */
-    public int getLevel() {
-        return level;
-    }
-
-    void setLevel(final int level) {
-        this.level = level;
-    }
-
-    /**
-     * @return optional message
-     */
-    public String getMessage() {
-        return message;
-    }
-
-    void setMessage(final String message) {
-        this.message = message;
-    }
-
-    public Date getDateTime() {
-        return dateTime;
-    }
-
-    void setDateTime(final Date dateTime) {
-        this.dateTime = dateTime;
-    }
-
-
-    public static Builder builder() {
-        return new Builder<>();
-    }
-
-    public static class Builder<S extends UnitStatus, B extends UnitStatus.Builder<S, B>> extends BaseStatus.Builder<S, B> {
-        private String unit;
-        private int level;
-        private String message;
-
-        public S build() {
-            final S status = super.build();
-            status.setUnit(unit);
-            status.setLevel(level);
-            status.setMessage(message);
-            status.setDateTime(Date.from(Instant.now()));
-            return status;
-        }
-
-        public B unit(final String unit) {
-            this.unit = unit;
-            return (B) this;
-        }
-
-        public B level(final int level) {
-            this.level = level;
-            return (B) this;
-        }
-
-        public B message(final String message) {
-            this.message = message;
-            return (B) this;
-        }
-
-        @Override
-        protected S createStatus() {
-            return (S) new UnitStatus();
-        }
-
-        @Override
-        public B rating(final Rating rating) {
-            return super.rating(rating);
-        }
-
-        @Override
-        public B worstRating(final Iterable<? extends BaseStatus> statuses) {
-            return super.worstRating(statuses);
-        }
-    }
-}
