--- common/src/test/java/org/opentestsystem/rdw/ingest/common/status/DiagnosticLevelTest.java
@@ -1,15 +1,0 @@
-package org.opentestsystem.rdw.ingest.common.status;
-
-import org.junit.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-public class DiagnosticLevelTest {
-
-    @Test
-    public void itShouldCompareAgainstLevel() {
-        assertThat(DiagnosticLevel.ReadDatabase.isLessThanOrEqualTo(1)).isFalse();
-        assertThat(DiagnosticLevel.ReadDatabase.isLessThanOrEqualTo(3)).isTrue();
-        assertThat(DiagnosticLevel.ReadDatabase.isLessThanOrEqualTo(5)).isTrue();
-    }
-}
