--- common/src/test/java/org/opentestsystem/rdw/ingest/common/status/LocalSystemStatusIndicatorTest.java
@@ -1,44 +1,0 @@
-package org.opentestsystem.rdw.ingest.common.status;
-
-import org.junit.Before;
-import org.junit.Test;
-import org.springframework.test.util.ReflectionTestUtils;
-
-import java.util.Map;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-/**
- * Note that this is meant as a pure unit test of {@link LocalSystemStatusIndicator}.
- * Spring magic wiring isn't dealt with here.
- */
-public class LocalSystemStatusIndicatorTest {
-
-    private LocalSystemStatusIndicator statusIndicator;
-
-    @Before
-    public void createStatusIndicator() {
-        statusIndicator = new LocalSystemStatusIndicator();
-        ReflectionTestUtils.setField(statusIndicator, "buildVersion", "1.2.3");
-    }
-
-    @Test
-    public void itShouldNotProduceStatusForLevel0() {
-        assertThat(statusIndicator.status(0)).isNull();
-    }
-
-    @Test
-    public void itShouldProduceStatusForLevel1() {
-        final Status status = statusIndicator.status(1);
-
-        // spot check a couple values
-        assertThat(status.getStatusRating()).isEqualTo(Rating.Ideal.value());
-
-        assertThat(status.getDetails().get("memory")).isNotNull();
-        assertThat(status.getDetails().get("processor")).isNotNull();
-        assertThat(status.getDetails().get("uptime")).isNotNull();
-
-        assertThat(((Map<String, Object>) status.getDetails().get("applicationInfo")).get("version")).isEqualTo("1.2.3");
-        assertThat(((Map<String, Object>) status.getDetails().get("memory")).get("jvmFreeMemory")).isNotNull();
-    }
-}
