--- common/src/test/java/org/opentestsystem/rdw/ingest/common/status/StatusTest.java
@@ -1,27 +1,0 @@
-package org.opentestsystem.rdw.ingest.common.status;
-
-import org.junit.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-public class StatusTest {
-
-    @Test
-    public void itHasDefault() {
-        final Status status = Status.builder().build();
-        assertThat(status.getStatusRating()).isEqualTo(4);
-        assertThat(status.getStatusText()).isEqualTo("Ideal");
-        assertThat(status.getDetails()).isEmpty();
-    }
-
-    @Test
-    public void itShouldPreserveBuilderValues() {
-        final Status status = Status.builder()
-                .rating(Rating.Degraded)
-                .detail("sub", "test")
-                .build();
-        assertThat(status.getStatusRating()).isEqualTo(Rating.Degraded.value());
-        assertThat(status.getStatusText()).isEqualTo(Rating.Degraded.name());
-        assertThat(status.getDetails().get("sub")).isEqualTo("test");
-    }
-}
