--- common/src/test/java/org/opentestsystem/rdw/ingest/common/status/SystemStatusTest.java
@@ -1,61 +1,0 @@
-package org.opentestsystem.rdw.ingest.common.status;
-
-import org.junit.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-public class SystemStatusTest {
-
-    @Test
-    public void itHasDefaults() {
-        final SystemStatus status = SystemStatus.builder().build();
-        assertThat(status.getStatusRating()).isEqualTo(4);
-        assertThat(status.getStatusText()).isEqualTo("Ideal");
-        assertThat(status.getDateTime()).isNotNull();
-    }
-
-    @Test
-    public void itShouldPreserveBuilderValues() {
-        final Status configurationStatus = Status.builder().build();
-        final Status databaseStatus = Status.builder().build();
-
-        final SystemStatus status = SystemStatus.builder()
-                .rating(Rating.Degraded)
-                .level(5)
-                .unit("test")
-                .sub("configuration", configurationStatus)
-                .sub("database", databaseStatus)
-                .build();
-        assertThat(status.getStatusRating()).isEqualTo(1);
-        assertThat(status.getStatusText()).isEqualTo("Degraded");
-        assertThat(status.getLevel()).isEqualTo(5);
-        assertThat(status.getUnit()).isEqualTo("test");
-        assertThat(status.getSubs().get("configuration")).isEqualTo(configurationStatus);
-        assertThat(status.getSubs().get("database")).isEqualTo(databaseStatus);
-    }
-
-    @Test
-    public void itShouldInheritWorstRating() {
-        final Status configurationStatus = Status.builder().rating(Rating.Degraded).build();
-
-        final SystemStatus status = SystemStatus.builder()
-                .level(1)
-                .unit("test")
-                .sub("configuration", configurationStatus)
-                .build();
-        assertThat(status.getStatusRating()).isEqualTo(Rating.Degraded.value());
-    }
-
-    @Test
-    public void itShouldNotInheritWorstRatingIfRatingDirectlySet() {
-        final Status configurationStatus = Status.builder().rating(Rating.Failed).build();
-
-        final SystemStatus status = SystemStatus.builder()
-                .rating(Rating.Warning)
-                .level(1)
-                .unit("test")
-                .sub("configuration", configurationStatus)
-                .build();
-        assertThat(status.getStatusRating()).isEqualTo(Rating.Warning.value());
-    }
-}
