+++ import-service/src/main/java/org/opentestsystem/rdw/ingest/ImportServiceApplication.java
@@ -1,7 +1,7 @@
 package org.opentestsystem.rdw.ingest;
 
 import org.opentestsystem.rdw.archive.ArchiveConfiguration;
-import org.opentestsystem.rdw.ingest.common.status.StatusConfiguration;
+import org.opentestsystem.rdw.common.status.StatusConfiguration;
 import org.opentestsystem.rdw.utils.YamlPropertiesConfigurator;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
