+++ import-service/src/main/java/org/opentestsystem/rdw/ingest/status/DatabaseStatusIndicator.java
@@ -1,14 +1,14 @@
 package org.opentestsystem.rdw.ingest.status;
 
 import org.opentestsystem.rdw.archive.ArchiveService;
+import org.opentestsystem.rdw.common.status.AbstractStatusIndicator;
+import org.opentestsystem.rdw.common.status.DiagnosticLevel;
+import org.opentestsystem.rdw.common.status.Rating;
+import org.opentestsystem.rdw.common.status.Status;
+import org.opentestsystem.rdw.common.status.StatusIndicator;
 import org.opentestsystem.rdw.ingest.common.model.ImportContent;
 import org.opentestsystem.rdw.ingest.common.model.ImportStatus;
 import org.opentestsystem.rdw.ingest.common.model.RdwImport;
-import org.opentestsystem.rdw.ingest.common.status.AbstractStatusIndicator;
-import org.opentestsystem.rdw.ingest.common.status.DiagnosticLevel;
-import org.opentestsystem.rdw.ingest.common.status.Rating;
-import org.opentestsystem.rdw.ingest.common.status.Status;
-import org.opentestsystem.rdw.ingest.common.status.StatusIndicator;
 import org.opentestsystem.rdw.ingest.repository.RdwImportRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
