+++ import-service/src/main/java/org/opentestsystem/rdw/ingest/status/ProvidersIndicator.java
@@ -1,9 +1,10 @@
 package org.opentestsystem.rdw.ingest.status;
 
-import org.opentestsystem.rdw.ingest.common.status.AbstractStatusIndicator;
-import org.opentestsystem.rdw.ingest.common.status.DiagnosticLevel;
-import org.opentestsystem.rdw.ingest.common.status.Status;
-import org.opentestsystem.rdw.ingest.common.status.UnitStatus;
+import org.opentestsystem.rdw.common.status.AbstractStatusIndicator;
+import org.opentestsystem.rdw.common.status.DiagnosticLevel;
+import org.opentestsystem.rdw.common.status.Status;
+import org.opentestsystem.rdw.common.status.StatusIndicator;
+import org.opentestsystem.rdw.common.status.UnitStatus;
 import org.springframework.stereotype.Component;
 
 import static com.google.common.collect.Lists.newArrayList;
