+++ import-service/src/test/java/org/opentestsystem/rdw/ingest/status/DatabaseStatusIndicatorTest.java
@@ -3,9 +3,9 @@ package org.opentestsystem.rdw.ingest.status;
 import org.junit.Before;
 import org.junit.Test;
 import org.opentestsystem.rdw.archive.ArchiveService;
+import org.opentestsystem.rdw.common.status.Status;
 import org.opentestsystem.rdw.ingest.common.model.ImportStatus;
 import org.opentestsystem.rdw.ingest.common.model.RdwImport;
-import org.opentestsystem.rdw.ingest.common.status.Status;
 import org.opentestsystem.rdw.ingest.repository.RdwImportRepository;
 
 import java.util.Arrays;
