+++ migrate-reporting/src/main/java/org/opentestsystem/rdw/ingest/migrate/reporting/MigrateReportingApplication.java
@@ -1,6 +1,6 @@
 package org.opentestsystem.rdw.ingest.migrate.reporting;
 
-import org.opentestsystem.rdw.ingest.common.status.StatusConfiguration;
+import org.opentestsystem.rdw.common.status.StatusConfiguration;
 import org.opentestsystem.rdw.utils.YamlPropertiesConfigurator;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
