+++ migrate-reporting/src/main/java/org/opentestsystem/rdw/ingest/migrate/reporting/status/DatabaseStatusIndicator.java
@@ -1,9 +1,9 @@
 package org.opentestsystem.rdw.ingest.migrate.reporting.status;
 
-import org.opentestsystem.rdw.ingest.common.status.AbstractStatusIndicator;
-import org.opentestsystem.rdw.ingest.common.status.DiagnosticLevel;
-import org.opentestsystem.rdw.ingest.common.status.Status;
-import org.opentestsystem.rdw.ingest.common.status.StatusIndicator;
+import org.opentestsystem.rdw.common.status.AbstractStatusIndicator;
+import org.opentestsystem.rdw.common.status.DiagnosticLevel;
+import org.opentestsystem.rdw.common.status.Status;
+import org.opentestsystem.rdw.common.status.StatusIndicator;
 import org.opentestsystem.rdw.ingest.migrate.reporting.Migrate;
 import org.opentestsystem.rdw.ingest.migrate.reporting.MigrateStatus;
 import org.opentestsystem.rdw.ingest.migrate.reporting.repository.ReportingMigrateRepository;
