+++ migrate-reporting/src/main/java/org/opentestsystem/rdw/ingest/migrate/reporting/status/MigrateReportingJobStatusIndicator.java
@@ -1,10 +1,10 @@
 package org.opentestsystem.rdw.ingest.migrate.reporting.status;
 
-import org.opentestsystem.rdw.ingest.common.status.AbstractStatusIndicator;
-import org.opentestsystem.rdw.ingest.common.status.DiagnosticLevel;
-import org.opentestsystem.rdw.ingest.common.status.Rating;
-import org.opentestsystem.rdw.ingest.common.status.Status;
-import org.opentestsystem.rdw.ingest.common.status.StatusIndicator;
+import org.opentestsystem.rdw.common.status.AbstractStatusIndicator;
+import org.opentestsystem.rdw.common.status.DiagnosticLevel;
+import org.opentestsystem.rdw.common.status.Rating;
+import org.opentestsystem.rdw.common.status.Status;
+import org.opentestsystem.rdw.common.status.StatusIndicator;
 import org.opentestsystem.rdw.ingest.migrate.reporting.MigrateLifecycle;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
