+++ migrate-reporting/src/test/java/org/opentestsystem/rdw/ingest/migrate/reporting/status/MigrateReportingJobStatusIndicatorTest.java
@@ -2,8 +2,8 @@ package org.opentestsystem.rdw.ingest.migrate.reporting.status;
 
 import org.junit.Before;
 import org.junit.Test;
-import org.opentestsystem.rdw.ingest.common.status.Rating;
-import org.opentestsystem.rdw.ingest.common.status.Status;
+import org.opentestsystem.rdw.common.status.Rating;
+import org.opentestsystem.rdw.common.status.Status;
 import org.opentestsystem.rdw.ingest.migrate.reporting.MigrateLifecycle;
 
 import static org.assertj.core.api.Assertions.assertThat;
