+++ task-service/build.gradle
@@ -7,7 +7,6 @@ dependencies {
     // handy for creating metadata with @ConfigurationProperties
     optional "org.springframework.boot:spring-boot-configuration-processor"
 
-    compile 'org.springframework.boot:spring-boot-starter-actuator'
     compile 'org.springframework.cloud:spring-cloud-starter-config'
     compile 'org.opentestsystem.rdw.common:rdw-common-utils'
 
@@ -28,7 +27,7 @@ dependencies {
 
     compile 'mysql:mysql-connector-java'
 
-    // used for status endpoint
+    compile 'org.opentestsystem.rdw.common:rdw-common-status'
     compile project(':rdw-ingest-common')
 
     testCompile 'org.springframework.boot:spring-boot-starter-test'
