+++ task-service/src/main/java/org/opentestsystem/rdw/ingest/tasking/TaskServiceApplication.java
@@ -1,7 +1,7 @@
 package org.opentestsystem.rdw.ingest.tasking;
 
 import org.opentestsystem.rdw.utils.YamlPropertiesConfigurator;
-import org.opentestsystem.rdw.ingest.common.status.StatusConfiguration;
+import org.opentestsystem.rdw.common.status.StatusConfiguration;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Import;
