+++ task-service/src/main/java/org/opentestsystem/rdw/ingest/tasking/service/ImportServiceClient.java
@@ -1,6 +1,6 @@
 package org.opentestsystem.rdw.ingest.tasking.service;
 
-import org.opentestsystem.rdw.ingest.common.status.Rating;
+import org.opentestsystem.rdw.common.status.Rating;
 
 /**
  * An import-service client.
