+++ task-service/src/main/java/org/opentestsystem/rdw/ingest/tasking/service/impl/RestImportServiceClient.java
@@ -1,8 +1,8 @@
 package org.opentestsystem.rdw.ingest.tasking.service.impl;
 
+import org.opentestsystem.rdw.common.status.Rating;
+import org.opentestsystem.rdw.common.status.UnitStatus;
 import org.opentestsystem.rdw.ingest.common.model.RdwImport;
-import org.opentestsystem.rdw.ingest.common.status.Rating;
-import org.opentestsystem.rdw.ingest.common.status.UnitStatus;
 import org.opentestsystem.rdw.ingest.tasking.service.ImportServiceClient;
 import org.opentestsystem.rdw.ingest.tasking.task.UpdateOrganizationsConfiguration;
 import org.slf4j.Logger;
