+++ task-service/src/main/java/org/opentestsystem/rdw/ingest/tasking/status/ReconciliationReportStatusIndicator.java
@@ -1,11 +1,11 @@
 package org.opentestsystem.rdw.ingest.tasking.status;
 
-import org.opentestsystem.rdw.ingest.common.status.AbstractStatusIndicator;
-import org.opentestsystem.rdw.ingest.common.status.DiagnosticLevel;
-import org.opentestsystem.rdw.ingest.common.status.Rating;
-import org.opentestsystem.rdw.ingest.common.status.Status;
-import org.opentestsystem.rdw.ingest.common.status.StatusIndicator;
-import org.opentestsystem.rdw.ingest.common.status.UnitStatus;
+import org.opentestsystem.rdw.common.status.AbstractStatusIndicator;
+import org.opentestsystem.rdw.common.status.DiagnosticLevel;
+import org.opentestsystem.rdw.common.status.Rating;
+import org.opentestsystem.rdw.common.status.Status;
+import org.opentestsystem.rdw.common.status.StatusIndicator;
+import org.opentestsystem.rdw.common.status.UnitStatus;
 import org.opentestsystem.rdw.ingest.tasking.repository.ReportRepository;
 import org.opentestsystem.rdw.ingest.tasking.service.impl.LoggingJavaMailSender;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,7 +14,6 @@ import org.springframework.mail.javamail.JavaMailSenderImpl;
 import org.springframework.stereotype.Component;
 
 import javax.mail.MessagingException;
-
 import java.util.ArrayList;
 
 import static com.google.common.collect.Lists.newArrayList;
