+++ task-service/src/main/java/org/opentestsystem/rdw/ingest/tasking/status/UpdateOrganizationsStatusIndicator.java
@@ -1,11 +1,11 @@
 package org.opentestsystem.rdw.ingest.tasking.status;
 
-import org.opentestsystem.rdw.ingest.common.status.AbstractStatusIndicator;
-import org.opentestsystem.rdw.ingest.common.status.DiagnosticLevel;
-import org.opentestsystem.rdw.ingest.common.status.Rating;
-import org.opentestsystem.rdw.ingest.common.status.Status;
-import org.opentestsystem.rdw.ingest.common.status.StatusIndicator;
-import org.opentestsystem.rdw.ingest.common.status.UnitStatus;
+import org.opentestsystem.rdw.common.status.AbstractStatusIndicator;
+import org.opentestsystem.rdw.common.status.DiagnosticLevel;
+import org.opentestsystem.rdw.common.status.Rating;
+import org.opentestsystem.rdw.common.status.Status;
+import org.opentestsystem.rdw.common.status.StatusIndicator;
+import org.opentestsystem.rdw.common.status.UnitStatus;
 import org.opentestsystem.rdw.ingest.tasking.repository.OrganizationRepository;
 import org.opentestsystem.rdw.ingest.tasking.service.ImportServiceClient;
 import org.opentestsystem.rdw.ingest.tasking.task.UpdateOrganizationsConfiguration;
