+++ task-service/src/test/java/org/opentestsystem/rdw/ingest/tasking/status/UpdateOrganizationsStatusIndicatorTest.java
@@ -2,9 +2,9 @@ package org.opentestsystem.rdw.ingest.tasking.status;
 
 import org.junit.Before;
 import org.junit.Test;
-import org.opentestsystem.rdw.ingest.common.status.Rating;
-import org.opentestsystem.rdw.ingest.common.status.Status;
-import org.opentestsystem.rdw.ingest.common.status.UnitStatus;
+import org.opentestsystem.rdw.common.status.Rating;
+import org.opentestsystem.rdw.common.status.Status;
+import org.opentestsystem.rdw.common.status.UnitStatus;
 import org.opentestsystem.rdw.ingest.tasking.repository.OrganizationRepository;
 import org.opentestsystem.rdw.ingest.tasking.service.ImportServiceClient;
